% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{computeMonteCarloPmc}
\alias{computeMonteCarloPmc}
\title{Monte Carlo \eqn{P_{\rm{mc}}} computation}
\usage{
computeMonteCarloPmc(
  paramsList,
  mcSamples = 1e+05,
  batchSize = mcSamples,
  numCores = 1,
  verbose = T
)
}
\arguments{
\item{paramsList}{List containing lists with each component GMM parameters. See \code{generateDistbnFunc} for format of components.}

\item{mcSamples}{Numeric for number of MC samples to use to approximate the integral. Default 1e5}

\item{batchSize}{Numeric for the observations to assign to each core. Helps with memory concerns. Default 1e3.}

\item{numCores}{Number of cores to use in parallel::mclapply call. Default is \code{floor(detectCores() / 2)}.}

\item{verbose}{Boolean whether to print output messages}
}
\value{
Value of Pmc from the Monte Carlo integral
}
\description{
Compute Monte Carlo estimate of Pmc associated with given cluster parameters based on Gaussian cluster assumption.
}
\details{
TODO: Fill me in.
}
