% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{computePairwisePmcMatrix}
\alias{computePairwisePmcMatrix}
\title{Pairwise \eqn{P_{\rm{mc}}} Matrix computation}
\usage{
computePairwisePmcMatrix(
  paramsList,
  mcSamples,
  numCores = 1,
  equalProbs = F,
  threshold = 0,
  verbose = T
)
}
\arguments{
\item{paramsList}{List containing lists with each component GMM parameters. See \code{generateDistbnFunc} for format of components.}

\item{mcSamples}{Monte Carlo Samples for Pmc calculation}

\item{numCores}{Number of cores for computations}

\item{equalProbs}{Whether to ignore prior weights for each component (i.e., Pmc based solely on mean/variance)}

\item{threshold}{Pmc threshold for Pairwise Pmc, all values below threshold are treated as 0}

\item{verbose}{Whether to print progress messages}

\item{...}{Additional parameters passed to \code{\link[=computePmc]{computePmc()}} or \code{\link[=computeMonteCarloPmc]{computeMonteCarloPmc()}}}
}
\value{
\eqn{K \times K} matrix with Pairwise \eqn{P_{\rm{mc}}} values for each pair of clusters
}
\description{
TODO: FILL ME IN
}
