% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{computePairwisePmcMatrixOLD}
\alias{computePairwisePmcMatrixOLD}
\title{Pairwise \eqn{P_{\rm{mc}}} Matrix computation}
\usage{
computePairwisePmcMatrixOLD(paramsList, mc = T, ...)
}
\arguments{
\item{paramsList}{List containing lists with each component GMM parameters. See \code{generateDistbnFunc} for format of components.}

\item{mc}{Boolean whether to compute \eqn{P_{\rm mc}} with Monte Carlo integration or cubature integration}

\item{...}{Additional parameters passed to \code{\link[=computePmc]{computePmc()}} or \code{\link[=computeMonteCarloPmc]{computeMonteCarloPmc()}}}
}
\value{
\eqn{K \times K} matrix with Pairwise \eqn{P_{\rm{mc}}} values for each pair of clusters
}
\description{
TODO: FILL ME IN
}
