% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{computePmc}
\alias{computePmc}
\title{Cubature \eqn{P_{\rm{mc}}} computation}
\usage{
computePmc(paramsList, integralControl = list())
}
\arguments{
\item{paramsList}{List containing lists with each component GMM parameters. See \code{generateDistbnFunc} for format of components.}

\item{integralControl}{List specifying arguments to pass to \code{\link[cubature:cubintegrate]{cubature::cubintegrate()}}. See details.}
}
\value{
Output from the \code{cubature::cubintegrate()} function.
}
\description{
Compute Pmc for a given Gaussian mixture distribution based on \code{cubature} package
}
\details{
For a given cluster configuration, the overall misclassification probability \eqn{P_{\rm mc}} can be evaluated as
\deqn{P_{\rm mc} = \sum_{j=1}^K \int 2 \left(1 - \pi_j(x)\right) \, \pi_j(x) P(x) dx}

This integral is implemented using the \code{cubature} function. the \code{integralControl} variable accepts arguments to the \code{\link[cubature:cubintegrate]{cubature::cubintegrate()}} function
The defaults for this function are:
\itemize{
\item \code{method}: Which integration method is to be used. Default is \code{hcubature}
\item \code{lowerLimit} and \code{upperLimit}: The bounds of integration. Default is \eqn{\pm \infty}.
\item \code{maxEval}: Sets the maximum number of integral evaluations. Default is 1e6
\item \code{relTol}: Sets the convergence tolerance for the integration. Default is 1e-5
}
}
\examples{
set.seed(1)
dat <- matrix(c(rnorm(200), rnorm(200, 3), rnorm(200, -3)), ncol=2, byrow=T)
partition <- c(rep(1, 100), rep(2, 100), rep(3, 100))
params <- constructPmcParamsPartition(partition, dat, G=1:5)
computePmc(params)

}
