% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmc_helpers.R
\name{computePosteriorProbMatrix}
\alias{computePosteriorProbMatrix}
\title{Compute Posterior Matrix for a set of parameteres}
\usage{
computePosteriorProbMatrix(paramsList, data)
}
\arguments{
\item{paramsList}{List of mixture parameteres}

\item{data}{Data matrix for which to compute posterior probabilities. Rows are observations}
}
\value{
An \eqn{N\times K} matrix
export
}
\description{
Compute the posterior probability matrix
}
