% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{consolidateParams}
\alias{consolidateParams}
\title{Consolidate redundant components}
\usage{
consolidateParams(paramsList, pmcMatrix, threshold)
}
\arguments{
\item{paramsList}{List containing lists with each component GMM parameters. See \code{generateDistbnFunc} for format of components.}

\item{pmcMatrix}{Pairwise Pmc matrix}

\item{threshold}{Pmc threshold for component consolidation. Any pairs of components with Pmc > threshold will be combined.}
}
\value{
List of parameters with redundant components consolidated inso a single component.
}
\description{
TODO: FILL ME IN
}
