% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{constructPmcParamsLocalizedEnsemble}
\alias{constructPmcParamsLocalizedEnsemble}
\title{E}
\usage{
constructPmcParamsLocalizedEnsemble(
  data,
  clustFunc,
  replicates,
  subsampSize = NULL,
  G = NULL,
  saveDir = NULL,
  prefix = "localizedEnsemble_",
  weighted = T,
  verbose = F,
  numCores = 1,
  seeds = NULL,
  ...
)
}
\description{
E
}
