% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{constructPmcParamsMclust}
\alias{constructPmcParamsMclust}
\title{Construct \eqn{P_{\rm{mc}}} parameter list from \code{Mclust} output}
\usage{
constructPmcParamsMclust(mclustObj, singleElement = F)
}
\arguments{
\item{mclustObj}{Output of \code{Mclust()} function call}

\item{singleElement}{Boolean, whether to combine into a single list element}
}
\value{
List of lists where each sublist contains the parameters for the
mixture component or a single list if \code{singleEleemnt} is \code{TRUE}.
}
\description{
Take the output of \code{Mclust()} and format it for consumption by \code{compute*Pmc} and \code{\link[=PHM]{PHM()}} functions
}
\details{
This function takes the parameters object from the output of \code{Mclust()} and
transforms it to define a GMM for use in the \code{compute*Pmc} and \code{PHM} functions. The output
is a list of lists, where each sublist corresponds ot the parameters of a single
cluster distribution (default is a single Gaussian component, K = 1).

If \code{singleElement = TRUE}, then all components will be combined into a single
list.

The parameters in each sublist are
\itemize{
\item \code{mean} should be a \code{DxK} matrix where each column corresponds to a component mean
\item \code{var} should be a \code{DxDxK} array where each slice corresponds to a covariance matrix
\item \code{prob} should be a \code{K} dimensional vector for the proportions within the parent mixture
}
}
\examples{
set.seed(1)
dat <- matrix(c(rnorm(200), rnorm(200, 3)), ncol=2, byrow=T)
mcl <- Mclust(dat)
constructPmcParamsMclust(mcl)

}
