% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{constructPmcParamsPartition}
\alias{constructPmcParamsPartition}
\title{Construct \eqn{P_{\rm{mc}}} parameter list from a partition}
\usage{
constructPmcParamsPartition(partition, data, ...)
}
\arguments{
\item{partition}{Vector of labels for observations}

\item{data}{Numeric matrix}
}
\value{
List of lists where each sublist contains the GMM density estimates
for each group in the partition.
}
\description{
Estimates the mixture model density for a partition of the data using \code{Mclust}
}
\details{
TODO: Fill me in.

See \code{\link[=constructPmcParamsMclust]{constructPmcParamsMclust()}} for a description of the output.
}
\examples{
# dat <- c(rnorm(100), rnorm(100, 3))
# partition <- c(rep(1, 100), rep(2, 100))
# constructPmcParamsPartition(partition, dat)

}
