% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{constructPmcParamsSubAggMclust}
\alias{constructPmcParamsSubAggMclust}
\title{Ensemble density estimation via Mclust}
\usage{
constructPmcParamsSubAggMclust(
  data,
  replicates,
  subsampSize,
  G = NULL,
  saveDir = NULL,
  prefix = "subAggMclust_",
  verbose = F,
  numCores = 1,
  ...
)
}
\arguments{
\item{data}{Numeric matrix for data}

\item{replicates}{Number of replicates over which to ensemble}

\item{G}{Number of components for each ensembled model}

\item{saveDir}{Where to save intermediate density estimates (for rerunning analyses). NULL is not saving}

\item{prefix}{Filename prefix for saving intermediate density estimates}

\item{verbose}{Whether to print out logging messages}

\item{numCores}{Number of cores over which to parallelize}

\item{...}{Parameters to provide to Mclust}

\item{modelNames}{Covariance structures for each ensembled moddel}

\item{subSampSize}{Number of subsamples to draw in each replicate}
}
\value{
List of lists where each sublist contains the
proportion, mean, covariance matrix estimates
for each component in the ensembled model.
}
\description{
EEP
}
\details{
FILL ME IN
}
\examples{
set seed(1)

}
