% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{constructPmcParamsSubAggPartition}
\alias{constructPmcParamsSubAggPartition}
\title{Ensemble weighted density estimation for partitions}
\usage{
constructPmcParamsSubAggPartition(
  data,
  partition,
  replicates,
  subsampSize,
  linkFunc = min,
  threshold = 1e-04,
  saveDir = NULL,
  prefix = "subAggPartition_",
  numCores = 1,
  verbose = F,
  ...
)
}
\arguments{
\item{data}{Numeric matrix for data (\eqn{N \times p})}

\item{partition}{Vector of labels for observations}

\item{replicates}{Number of replicates over which to be ensembled}

\item{linkFunc}{Function to use to combine all distances within a cluster for weight calculation. Default is \code{min}}

\item{threshold}{Threshold past which to not include weights (for computational efficiency)}

\item{saveDir}{Where to save intermediate density estimates (for rerunning analyses). NULL is not saving}

\item{prefix}{Filename prefix for saving intermediate density estimates}

\item{numCores}{Number of cores over which to parallelize}

\item{verbose}{Whether to print out logging messages}

\item{...}{Parameters passed to \code{\link[mclust:Mclust]{mclust::Mclust()}}}
}
\value{
List of lists where each sublist contains the
proportion, mean, covariance matrix estimates
for each cluster.
}
\description{
EEP
}
\details{
PEW
}
\examples{
set seed(1)

}
