% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altPHM.R
\name{convertToPHMv2}
\alias{convertToPHMv2}
\title{Convert an original PHM object to the new PHM format}
\usage{
convertToPHMv2(phm)
}
\arguments{
\item{phm}{Output from \code{\link[=PHM]{PHM()}} function}
}
\description{
Convert an original PHM object to the new PHM format
}
