% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{decomposeParams}
\alias{decomposeParams}
\title{Decompose cluster solutions}
\usage{
decomposeParams(paramsList)
}
\arguments{
\item{paramsList}{List containing lists with each component GMM parameters. See \code{generateDistbnFunc} for format of components.}
}
\value{
List of parameters with redundant components consolidated inso a single component.
}
\description{
TODO: FILL ME IN
}
