% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityEstimation.R
\name{distanceWeightFunc}
\alias{distanceWeightFunc}
\title{Distance-based Cluster Weight Calculation}
\usage{
distanceWeightFunc(
  X,
  partition,
  removeSelfDistance = F,
  scaling = 1,
  aggFunc = min
)
}
\arguments{
\item{X}{Data matrix}

\item{partition}{Partition of observations in X into clusters}

\item{removeSelfDistance}{Whether to count distance to self (i.e., distance to assigned cluster is 0)}

\item{scaling}{Factor by which to multiply distances. Larger values places more weight on the assigned cluster.}

\item{aggFunc}{Function to aggregate distances from the observation to the cluster. Default is min (minimum distance used).}
}
\description{
Constructs a weighting matrix based on the smallest distance to observations in a cluster.
}
\details{
FILL ME IN
}
