% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phraseDocFunctions.R
\name{freqPhrases}
\alias{freqPhrases}
\title{Display Frequent Principal Phrases}
\usage{
freqPhrases(pd, n = 10)
}
\arguments{
\item{pd}{A phraseDoc object.}

\item{n}{Number of principal phrases to display.}
}
\value{
A vector with the \code{n} most frequent principal phrases and their
        frequencies.
}
\description{
Display the most frequent principal phrases in a phraseDoc object.
}
\examples{
tst=c("This is a test text",
      "This is a test text 2",
      "This is another test text",
      "This is another test text 2",
      "This girl will test text that man",
      "This boy will test text that man")
pd=phraseDoc(tst)
freqPhrases(pd, 2)
}
