% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityEstimation.R
\name{globalDensityEstimation}
\alias{globalDensityEstimation}
\title{Global (Ensemble) Density Estimation}
\usage{
globalDensityEstimation(
  X,
  M = 1,
  rho = NULL,
  subsampleSize = NULL,
  replace = F,
  seeds = 1:M,
  numCores = 1,
  saveDir = NULL,
  prefix = "global",
  overwrite = F,
  verbose = F,
  ...
)
}
\arguments{
\item{X}{Numeric matrix for data}

\item{M}{Number of replicates over which to ensemble. Default is 1, producing a single density estimate.}

\item{rho}{Proportion of dataset for subsample. Default is NULL, using the entire dataset.}

\item{replace}{Whether to sample with replacement for subsampling.}

\item{seeds}{List of seeds to use across replicates.}

\item{numCores}{Number of cores over which to parallelize the density estimation.}

\item{saveDir}{Where to save intermediate density estimates (for rerunning analyses). NULL is not saving.}

\item{prefix}{Filename prefix for saving intermediate density estimates. Ignored if saveDir is NULL.}

\item{overwrite}{Whether to overwrite the results if they already exist. Ignoreed if saveDir is NULL}

\item{verbose}{Whether to print out logging messages.}

\item{...}{Parameters to provide to Mclust.}

\item{subSampSize}{Explicitly value for subsample size. Default is NULL, using the entire dataset.}
}
\value{
List of lists where each sublist contains the
proportion, mean, covariance matrix estimates
for each component in the ensembled model.
}
\description{
EEP
}
\details{
FILL ME IN
}
\examples{
set seed(1)

}
