% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityEstimation.R
\name{localizedDensityEstimation}
\alias{localizedDensityEstimation}
\title{Localized (Ensemble) Density Estimation}
\usage{
localizedDensityEstimation(
  X,
  partitionFunc = NULL,
  partitionFuncParams = list(),
  weightFunc = NULL,
  weightFuncParams = list(),
  weightThreshold = 0.005,
  preserveClusters = F,
  M = 1,
  rho = NULL,
  subsampleSize = NULL,
  replace = F,
  seeds = 1:M,
  numCores = 1,
  saveDir = NULL,
  prefix = "global",
  overwrite = F,
  verbose = F,
  ...
)
}
\arguments{
\item{X}{Numeric matrix for data}

\item{partitionFunc}{Function taking input matrix X as input and returning a vector of partition labels.}

\item{partitionFuncParams}{List of parameters to be passed to partitionFunc.}

\item{weightFunc}{Function that acceps input matrix X and cluster vector as inputs, returns an NxK weight matrix.}

\item{weightFuncParams}{List of parameters to be passed to weightFunc.}

\item{weightThreshold}{For weighted GMM, ignore observations with weight below this value.}

\item{preserveClusters}{Whether to keep the components for a partition together or to treat them separately}

\item{M}{Number of replicates over which to ensemble. Default is 1, producing a single density estimate.}

\item{rho}{Proportion of dataset for subsample. Default is NULL, using the entire dataset.}

\item{replace}{Whether to sample with replacement for subsampling.}

\item{seeds}{List of seeds to use across replicates.}

\item{numCores}{Number of cores over which to parallelize the density estimation.}

\item{saveDir}{Where to save intermediate density estimates (for rerunning analyses). NULL is not saving.}

\item{prefix}{Filename prefix for saving intermediate density estimates. Ignored if saveDir is NULL.}

\item{overwrite}{Whether to overwrite the results if they already exist. Ignoreed if saveDir is NULL}

\item{verbose}{Whether to print out logging messages.}

\item{...}{Parameters to provide to Mclust.}
}
\description{
EEP
}
\details{
FILL ME IN
}
