% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityEstimation.R
\name{naiveWeightFunc}
\alias{naiveWeightFunc}
\title{Naive Cluster Weight Calculation}
\usage{
naiveWeightFunc(X, partition)
}
\arguments{
\item{X}{Data matrix}

\item{partition}{Partition of observations in X into clusters}
}
\description{
Constructs a one-hot weight matrix based on cluster assignments.
}
\details{
An observation is assigned weight 1 for cluster k if it is a member of cluster k, 0 otherwise.
}
