% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHM.R
\name{PHM}
\alias{PHM}
\title{PHM Algorithm}
\usage{
PHM(
  mclustObj = NULL,
  paramsList = NULL,
  partition = NULL,
  data = NULL,
  verbose = T,
  computePosterior = T,
  partitionModel = "VVI",
  partitionMaxComponents = 10,
  mc = T,
  ...
)
}
\value{
FILL ME IN
}
\description{
TODO: FILL ME IN
}
