% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotPHMMatrix}
\alias{plotPHMMatrix}
\title{Visualize PHM dendrogram structure}
\usage{
plotPHMMatrix(
  phm,
  colors = NULL,
  displayAxis = c("box", "label", "index", "none"),
  displayAxisSize = NULL,
  colorAxis = NULL,
  gridColor = "black",
  legendPosition = "none"
)
}
\arguments{
\item{doot}{pew}
}
\value{
pew
}
\description{
TODO: Fill me in
}
\details{
FILL ME IN
}
