% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotPmc2D}
\alias{plotPmc2D}
\title{Visualize regions contributing to \eqn{P_{\rm mc}} in a 2D plot}
\usage{
plotPmc2D(
  paramsList,
  data,
  partition,
  colors = RColorBrewer::brewer.pal(length(paramsList), "Paired"),
  xlim = NULL,
  ylim = NULL,
  suppressPmc = F,
  numPmcPatches = 200,
  logPmcThreshold = -4,
  logPmcMidpoint = 0.6 * logPmcThreshold,
  PmcColor = "#222",
  suppressDensity = F,
  densityLevels = c(0.01, 0.1),
  densityLevelWidth = 0.3,
  suppressObservations = F,
  pointSize = 0.5,
  labelSize = 2,
  textSize = 9,
  legendPosition = "none"
)
}
\arguments{
\item{paramsList}{A list generated from \code{\link[=constructPmcParamsMclust]{constructPmcParamsMclust()}}, \code{\link[=constructPmcParamsPartition]{constructPmcParamsPartition()}}, \code{\link[=constructPmcParamsPartition]{constructPmcParamsPartition()}} providing the initial cluster parameter estimates. Parameters should be 2D}

\item{data}{A \eqn{N \times 2} matrix containing observations}

\item{partition}{A vector containing class labels}

\item{colors}{Vector of \eqn{K} hex codes to color the leaf node labels}

\item{xlim}{Vector with min/max values for the x-axis \code{NULL} sets this based on the maximum and minimum x-values for observations}

\item{ylim}{Vector with min/max values for the y-axis \code{NULL} sets this based on the maximum and minimum y-values for observations}

\item{suppressPmc}{Flag whether to display the regions contributing to \eqn{P_{\rm mc}} in the plot}

\item{numPmcPatches}{Number of points along each axis in the \eqn{P_{mc}} grid to evaluate. Higher values gives a smoother visualization}

\item{logPmcThreshold}{Threshold for \eqn{P_{\rm mc}} values to display. All coordinates with \eqn{\log_{10} P_{\rm mc}} below this value will be ignored.}

\item{logPmcMidpoint}{Midpoint in the \eqn{P_{\rm mc}} color gradient}

\item{PmcColor}{Hex value for the maximum value in the \eqn{P_{\rm mc}} color gradient}

\item{suppressDensity}{Flag whether to overlay the cluster-specific densities}

\item{densityLevels}{Values of the density at which to display the level curves}

\item{densityLevelWidth}{Line width for the density curves}

\item{suppressObservations}{Flag whether to display the observations from \code{data}}

\item{pointSize}{Size of the data scatterplot points}

\item{labelSize}{Text size for the cluster labels}

\item{textSize}{Text size for the plots}

\item{legendPosition}{Where to put the legend for the heatmap colors. Default is to suppress}
}
\value{
A ggplot object
}
\description{
Visualize the point-specific \eqn{P_{\rm mc}} over a grid of points to visually inspect cluster contributions to \eqn{P_{\rm mc}}
}
\details{
For an arbitrary point \eqn{\mathbf x}, using the Monte Carlo evaluation of \eqn{P_{\rm mc}} its point-specific \eqn{P_{\rm mc}} can be calculated as
\deqn{\sum_{k=1}^K \pi_k(\mathbf{x}) (1 - \pi_k(\mathbf{x})) \times f(x=\mathbf{x}) }
Where \eqn{f(\mathbf{x})} is the overall data density evaluated at a point. Note that for a sample of \eqn{M} points, the average of these point-specific \eqn{P_{\rm mc}} values produce the Monte Carlo estimate of \eqn{P_{\rm mc}} described in Turfah and Wen (2025).

The point-specific \eqn{P_{\rm mc}} of each point in a grid is evaluated and visualized. The cluster-specific density level sets and/or the partitioned observations can be overlaid over this to better understand how the \eqn{P_{\rm mc}} value was obtained.
}
