% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotPmcMatrix}
\alias{plotPmcMatrix}
\title{Plot \eqn{\Delta \P_{\rm{mc}}} matrix}
\usage{
plotPmcMatrix(
  phm,
  K = length(phm),
  colors = NULL,
  displayAxis = c("box", "label", "index", "none"),
  displayAxisSize = NULL,
  colorAxis = NULL,
  visScale = c("absolute", "percent"),
  visSize = 2,
  visThreshold = 0.001,
  visDigits = 3
)
}
\arguments{
\item{phm}{Output from \code{\link[=PHM]{PHM()}}}

\item{K}{Number of clusters for which to visualize the heatmap}

\item{colors}{Vector of \eqn{K} hex codes to color the leaf node labels}

\item{displayAxis}{String indicating what label to place along the axis (corresponding to clusters)}

\item{displayAxisSize}{Text size for the axis labels}

\item{colorAxis}{Whether or not to color the axis labels}

\item{visScale}{Whether to display the raw \eqn{\Delta P_{\rm{mc}}} values or scale them to be percent of total \eqn{P_{\rm{mc}}}}

\item{visSize}{Text size for the values inside the heatmap}

\item{visThreshold}{At what value suppress the value and show "< (visThreshold)"}

\item{visDigits}{Number of digits to round the displayed values}
}
\value{
A ggplot object
}
\description{
Visualize the matrix of \eqn{\Delta P_{\rm mc}} values
}
