% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phraseDocFunctions.R
\name{print.phraseDoc}
\alias{print.phraseDoc}
\title{Print a phraseDoc Object}
\usage{
\method{print}{phraseDoc}(x, ...)
}
\arguments{
\item{x}{Object of type phraseDoc}

\item{...}{Additional arguments}
}
\description{
Print a phraseDoc Object
}
\examples{
tst=c("This is a test text",
      "This is a test text 2",
      "This is another test text",
      "This is another test text 2",
      "This girl will test text that man",
      "This boy will test text that man")
(pd=phraseDoc(tst))
}
