% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{print.textCluster}
\alias{print.textCluster}
\title{Print a textCluster Object}
\usage{
\method{print}{textCluster}(x, ...)
}
\arguments{
\item{x}{Object of type textCluster}

\item{...}{Additional arguments}
}
\value{
The total number of clusters and total number of documents are 
printed. There is no return value.
}
\description{
Print a textCluster Object
}
\examples{
M=matrix(c(0,1,0,2,0,10,0,14,12,0,8,0,1,0,1,0),4)
colnames(M)=1:4;rownames(M)=c("A","B","C","D")
tc=textCluster(M,2)
tc
}
