% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altPHM.R
\name{recover...v2}
\alias{recover...v2}
\alias{recoverGroupsv2}
\alias{recoverPosteriorv2}
\alias{recoverDeltaPmcv2}
\alias{recoverParamsv2}
\title{Recover outputs for a given cut of the PHM tree}
\usage{
recoverGroupsv2(phm, k)

recoverPosteriorv2(phm, k, posterior)

recoverDeltaPmcv2(phm, k)

recoverParamsv2(phm, k, paramsToKeep = c("prob", "mean", "var", "class"))
}
\arguments{
\item{phm}{Output from the \code{\link[=PHMv2]{PHMv2()}} function.}

\item{k}{Number of clusters at which to recover the group structure}

\item{posterior}{Original posterior matrix from which to combine clusters}

\item{paramsToKeep}{Fields from \code{phm$paramsList} to preserve}
}
\description{
Recover outputs for a given cut of the PHM tree

Recover posterior at given value of \eqn{K}

Recover deltaPmc matrix at given value of \eqn{K}

Recover deltaPmc matrix at given value of \eqn{K}
}
