% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHM.R
\name{recoverDeltaPmcMatrix}
\alias{recoverDeltaPmcMatrix}
\title{Recover the posterior matrix based on PHM merging}
\usage{
recoverDeltaPmcMatrix(phm, K)
}
\arguments{
\item{phm}{Output from \code{\link[=PHM]{PHM()}}}

\item{K}{Number of clusters for which to compute the deltaPmc Matrix}
}
\value{
deltaPmc matrix for PHM at step K
}
\description{
Recover the posterior matrix based on PHM merging
}
\details{
TODO: Fill me in
}
