% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHM.R
\name{recoverPHMParams}
\alias{recoverPHMParams}
\title{Recover the parameters based on the PHM merging}
\usage{
recoverPHMParams(phm, K, paramsToKeep = c("prob", "mean", "var", "class"))
}
\arguments{
\item{phm}{Output from \code{\link[=PHM]{PHM()}}}

\item{K}{Number of clusters for which to compute the deltaPmc Matrix}

\item{paramsToKeep}{Parameters to preserve in merging.}
}
\value{
Parameter list.
}
\description{
Recover the parameters based on the PHM merging
}
\details{
TODO: Fill me in
}
