% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHM.R
\name{recoverPosterior}
\alias{recoverPosterior}
\title{Recover the posterior matrix based on the PHM merging}
\usage{
recoverPosterior(phm, K, data = NULL, computePosterior = F)
}
\arguments{
\item{phm}{Output from \code{\link[=PHM]{PHM()}}}

\item{K}{Number of clusters for which to compute the deltaPmc Matrix}

\item{data}{Matrix of observations for which to compute the posterior matrix}

\item{computePosterior}{Whether to recompute the posterior matrix if not available}
}
\value{
TODO
}
\description{
Recover the posterior matrix based on the PHM merging
}
\details{
TODO: Fill me in
}
