% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altPHM.R
\name{recoverPosteriorv2}
\alias{recoverPosteriorv2}
\title{Recover posterior at given value of \eqn{K}}
\usage{
recoverPosteriorv2(phm, k, posterior)
}
\description{
Recover posterior at given value of \eqn{K}
}
