% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phraseDocFunctions.R
\name{removePhrases}
\alias{removePhrases}
\title{Remove Phrases from phraseDoc Object}
\usage{
removePhrases(pd, phrs)
}
\arguments{
\item{pd}{A phraseDoc object.}

\item{phrs}{A set of phrases.}
}
\value{
A phraseDoc object with the phrases in \code{phrs} removed.
}
\description{
Remove a set of phrases from a phraseDoc object.
}
\examples{
tst=c("This is a test text",
      "This is a test text 2",
      "This is another test text",
      "This is another test text 2",
      "This girl will test text that man",
      "This boy will test text that man")
pd=phraseDoc(tst)
removePhrases(pd, c("test text","another test text"))
}
