% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phraseDocFunctions.R
\name{stopEndWords}
\alias{stopEndWords}
\title{Words that Principal Phrases do not End with}
\usage{
stopEndWords()
}
\value{
vector with words
}
\description{
Create a vector with words that principal phrases should not end with.
}
\examples{
stopEndWords()
}
