% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phraseDocFunctions.R
\name{stopPhrases}
\alias{stopPhrases}
\title{Phrases that are not Principal Phrases}
\usage{
stopPhrases()
}
\value{
vector with phrases
}
\description{
Create a vector with phrases that are not principal phrases.
}
\examples{
stopPhrases()
}
