% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phraseDocFunctions.R
\name{stopStartWords}
\alias{stopStartWords}
\title{Words that Principal Phrases do not Start with}
\usage{
stopStartWords()
}
\value{
vector with words
}
\description{
Create a vector with words that principal phrases should not start with.
}
\examples{
stopStartWords()
}
