% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHM.R
\name{thresholdPHM}
\alias{thresholdPHM}
\title{Find the solution for PHM at a specified threshold}
\usage{
thresholdPHM(phm_output, threshold = 0.01)
}
\arguments{
\item{phm_output}{Output of \code{PHM}}

\item{threshold}{Pmc threshold, default is 0.01

TODO: Fill in Returns + More detailed description}
}
\value{
Cluster results based on GMM satisfying specified threshold
}
\description{
Find the solution for PHM at a specified threshold
}
