% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHM.R
\name{addPosteriorMatrix}
\alias{addPosteriorMatrix}
\title{Compute Posterior Matrix based on PHM merging}
\usage{
addPosteriorMatrix(phm, data, initK = length(phm))
}
\description{
Compute Posterior Matrix based on PHM merging
}
\details{
TODO: Fill me in
}
