% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.phraseDoc.R
\name{as.matrix.phraseDoc}
\alias{as.matrix.phraseDoc}
\title{Convert a phraseDoc Object to a Matrix}
\usage{
\method{as.matrix}{phraseDoc}(x, ...)
}
\arguments{
\item{x}{A phraseDoc object.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
A matrix with phrases as rows, texts as columns, and elements
containing the number of times the phrase occurs in the text.
}
\description{
Convert a phraseDoc Object to a Matrix
}
\examples{
tst=c("This is a test text",
      "This is a test text 2",
      "This is another test text",
      "This is another test text 2",
      "This girl will test text that man",
      "This boy will test text that man")
pd=phraseDoc(tst)
as.matrix(pd)
}
