% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{computeDeltaPmcMatrix}
\alias{computeDeltaPmcMatrix}
\title{\eqn{\Delta P_{\rm {mc}}} Matrix computation}
\usage{
computeDeltaPmcMatrix(paramsList, integralControl = list())
}
\arguments{
\item{paramsList}{List containing lists with each component GMM parameters. See \code{generateDistbnFunc} for format of components.}

\item{integralControl}{List specifying arguments to pass to \code{\link[cubature:cubintegrate]{cubature::cubintegrate()}}. See details.}
}
\value{
\eqn{K \times K} matrix with each pair of clusters' contribution to \eqn{P_{\rm{mc}}}
}
\description{
Compute the merging Pmc reduction matrix calculated using \code{cubature} package
}
\details{
See \link{computePmc} for description of \code{integralControl} parameters.
}
