% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{computeDeltaPmcMatrix}
\alias{computeDeltaPmcMatrix}
\title{$\Delta P_{\rm {mc}}$ Matrix computation}
\usage{
computeDeltaPmcMatrix(paramsList, integralControl = list())
}
\description{
Compute the merging Pmc reduction matrix calculated using \code{cubature} package
}
\details{
See \code{\link[=computePmc]{computePmc()}} for a breakdown of the parameters
}
