% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{computeMonteCarloDeltaPmcMatrix}
\alias{computeMonteCarloDeltaPmcMatrix}
\title{Monte Carlo $\Delta P_{\rm {mc}}$ Matrix computation}
\usage{
computeMonteCarloDeltaPmcMatrix(
  paramsList,
  mcSamples = 1e+06,
  batchSize = mcSamples,
  numCores = 1,
  verbose = T
)
}
\description{
Compute the merging Pmc reduction matrix estimated via Monte Carlo integration
}
\details{
See \code{\link[=computeMonteCarloPmc]{computeMonteCarloPmc()}} for a breakdown of the parameters
}
