% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{computePmc}
\alias{computePmc}
\title{Cubature \eqn{P_{\rm{mc}}} computation}
\usage{
computePmc(paramsList, integralControl = list())
}
\arguments{
\item{paramsList}{List containing lists with each component GMM parameters. See \code{generateDistbnFunc} for format of components.}

\item{integralControl}{List specifying arguments to pass to \code{\link[cubature:cubintegrate]{cubature::cubintegrate()}}. See details.}
}
\value{
Output from the \code{cubature::cubintegrate()} function.
}
\description{
Compute Pmc for a given Gaussian mixture distribution based on \code{cubature} package
}
\details{
TODO: Clear this up
For \code{pcubature} the control variables are \code{method} for the integration
For \code{pcubature} the control variables are \code{lowerLimit} and \code{upperLimit} for the integral (defaults to \pm Inf)
For \code{pcubature} the control variables are \code{maxEval} which sets the number of integral evaluations (defaults to 1e6)
For \code{pcubature} the control variables are \code{relTol} which sets the maximum tolerance (defaults to 1e-5)
}
