% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmc.R
\name{constructPmcParamsPartition}
\alias{constructPmcParamsPartition}
\title{Construct \eqn{P_{\rm{mc}}} parameter list from a partition}
\usage{
constructPmcParamsPartition(partition, data, ...)
}
\arguments{
\item{partition}{Vector of labels for observations}

\item{data}{Numeric matrix}

\item{...}{Parameters passed to \code{\link[mclust:Mclust]{mclust::Mclust()}}}
}
\value{
List of lists where each sublist contains the
proportion, mean, covariance matrix estimates
for each cluster.
}
\description{
Estimates the mixture model density for a partition of the data using \code{Mclust}
}
\details{
Performs a naive density estimation, fitting a GMM to each partition separately.
Only observations in a cluster are considered to fit the GMM.

See \code{\link[=constructPmcParamsMclust]{constructPmcParamsMclust()}} for a description of the output.
}
\examples{
set.seed(1)
dat <- matrix(c(rnorm(200), rnorm(200, 3), rnorm(200, -3)), ncol=2, byrow=T)
partition <- c(rep(1, 100), rep(2, 100), rep(3, 100))
params <- constructPmcParamsPartition(partition, dat, G=1:5)

}
