% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altPHM.R
\name{PHMv2}
\alias{PHMv2}
\title{PHM Algorithm (supporting alternate merging strategies)}
\usage{
PHMv2(
  paramsList = NULL,
  deltaPmc = NULL,
  scaling = c("unscaled", "average", "alpha"),
  monteCarlo = T,
  numCores = 1,
  verbose = F,
  ...
)
}
\arguments{
\item{par1}{}
}
\value{
What is the return object
}
\description{
TODO: Fill me in
}
\examples{
set.seed(1)

}
