% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotPHMDendrogram}
\alias{plotPHMDendrogram}
\title{Visualize PHM merging procedure via Dendrogram}
\usage{
plotPHMDendrogram(
  phm,
  colors = NULL,
  uniformHeights = F,
  threshold = 0,
  suppress_labels = F,
  mergeLabels = c("delta", "pmc", "pct"),
  mergeLabelsSize = 2,
  mergeLabelsBorderSize = 0.15,
  mergeLabelsPadding = 0.15,
  mergeLabelsR = 0.1,
  displayAxis = c("box", "label", "index", "none"),
  displayAxisSize = NULL,
  colorAxis = NULL
)
}
\description{
Visualize PHM merging procedure via Dendrogram
}
\details{
TODO: Fill me in
}
