% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotPHMDendrogram}
\alias{plotPHMDendrogram}
\title{Visualize PHM merging procedure via Dendrogram}
\usage{
plotPHMDendrogram(
  phm,
  colors = NULL,
  scaleHeights = c("log10", "unscaled", "pmcdist"),
  heightValue = c("merge", "min"),
  threshold = 0,
  suppressLabels = F,
  mergeLabels = c("delta", "pmc", "percent"),
  mergeLabelsSize = 2,
  mergeLabelsBorderSize = 0.15,
  mergeLabelsPadding = 0.15,
  mergeLabelsR = 0.1,
  displayAxis = c("box", "label", "index", "none"),
  displayAxisSize = NULL,
  colorAxis = NULL,
  groupProbs = NULL,
  groupColorMax = "black",
  groupColorMin = "lightgray"
)
}
\arguments{
\item{phm}{Output from \code{\link[=PHM]{PHM()}}}

\item{colors}{Vector of \eqn{K} hex codes to color the leaf node labels}

\item{scaleHeights}{String specifying how to set the heights in the dendrogram. See Details for more information.}

\item{heightValue}{Whether to use the \eqn{\Delta P_{\rm mc}} or \eqn{\min \Delta P_{\rm mc}} value to determine branch height for a merge. See Details for more information.}

\item{threshold}{Error threshold for the integral past which to represent the merges as dashed lines}

\item{suppressLabels}{Boolean whether or not to display \eqn{P_{\rm{mc}}} reduction labels on the dendrogram or not}

\item{mergeLabels}{String indicating what value to display in the labels on the dendrogram}

\item{mergeLabelsSize}{Text size for the numeric value in the labels on the dendrogram}

\item{mergeLabelsPadding}{Padding for the merge label text}

\item{mergeLabelsR}{Radius for the rounded edges of the merge label box}

\item{displayAxis}{String indicating what label to place on the leaf nodes of the dendrogram. See Details for more information.}

\item{displayAxisSize}{Text size for the leaf node labels}

\item{colorAxis}{Whether or not to color the labels on the leaf nodes}

\item{groupProbs}{Vector of class probability conditional on base group membership}

\item{groupColorMax}{Color of lines corresponding to high group probability}

\item{groupColorMin}{Color of lines corresponding to low group probability}
}
\value{
A ggplot object
}
\description{
Visualize the PHM merging procedure using a dendrogram. Visualization options, such as displaying the merge \eqn{\Delta P_{\rm mc}} value and tracking group membership across merges is included.
}
\details{
There are two options for the value for the \eqn{P_{\rm mc}} height for merging subtrees \eqn{\mathcal{S}_1, \mathcal{S}_2} (\eqn{m = |\mathcal{S}_1| + |\mathcal{S}_2|}):
\itemize{
\item \code{"merge"}: \eqn{\binom{m}{2}^{-1} \Delta\pmis^{(\mathcal{S}_1, \mathcal{S}2)}}
\item \code{"min"}: \eqn{\min_{i \in \mathcal{S}_1,\, j \in \mathcal{S}_2} \Delta\pmis^{(i, j)}}
}

Once the value for the tree height is obtained, there are three options to scale.
\itemize{
\item \code{"unscaled"}: The height can be left unscaled
\item \code{"log10"}: A \eqn{\log_{10}} scaling can be applied to the height to better reveal different clustering resolutions.
\item \code{"pmcdist"}: A spline-based scaling where we map the value to a linear distance between two Gaussian clusters.
}

The \code{displayAxis} parameter controls what to display on the axes of the heatmap. \code{box} displays a standard box, which is most useful for color-coded axes. \code{label} displays the cluster label, as specified in the \code{phm} parameters. \code{index} displays the numeric index of each cluster in the \code{phm} parameter list. \code{none} suppresses the cluster label entirely, and is mose useful for when there are a large number of clusters.
}
