% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altPHM.R
\name{plotPHMv2Dendrogram}
\alias{plotPHMv2Dendrogram}
\title{Plot PHM Dendrogram (for v2)}
\usage{
plotPHMv2Dendrogram(
  phmObj,
  scaleHeights = c("log10", "unscaled", "pmcdist"),
  colors = NULL,
  displayAxis = c("box", "label", "index", "none"),
  displayAxisSize = NULL,
  colorAxis = NULL,
  groupProbs = NULL,
  groupColorMax = "black",
  groupColorMin = "lightgray"
)
}
\description{
Plot PHM Dendrogram (for v2)
}
