% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altPHM.R
\name{plotPHMv2Heatmap}
\alias{plotPHMv2Heatmap}
\title{Plot PHM Matrix (for v2)}
\usage{
plotPHMv2Heatmap(
  phmObj,
  colors = NULL,
  displayAxis = c("box", "label", "index", "none"),
  displayAxisSize = NULL,
  colorAxis = NULL,
  gridColor = "black",
  fillLimits = NULL,
  fillScale = c("log10", "pmcdist"),
  legendPosition = "none"
)
}
\description{
Plot PHM Matrix (for v2)
}
