% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotPmcMatrix}
\alias{plotPmcMatrix}
\title{Plot \eqn{\Delta \P_{\rm{mc}}} matrix}
\usage{
plotPmcMatrix(
  phm,
  K = length(phm),
  colors = NULL,
  displayAxis = c("box", "label", "index", "none"),
  displayAxisSize = NULL,
  colorAxis = NULL,
  visScale = c("absolute", "percent"),
  visSize = 2,
  visThreshold = 0.001,
  visDigits = 3
)
}
\arguments{
\item{temp}{doot}
}
\value{
pew
}
\description{
Plot \eqn{\Delta \P_{\rm{mc}}} matrix
}
