% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altPHM.R
\name{recoverParamsv2}
\alias{recoverParamsv2}
\title{Recover deltaPmc matrix at given value of \eqn{K}}
\usage{
recoverParamsv2(phm, k, paramsToKeep = c("prob", "mean", "var", "class"))
}
\description{
Recover deltaPmc matrix at given value of \eqn{K}
}
