% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHM.R
\name{thresholdPHM}
\alias{thresholdPHM}
\title{Find a clustering for a given \eqn{P_{\rm{mc}}} threshold}
\usage{
thresholdPHM(phm, threshold = 0.01)
}
\arguments{
\item{phm}{Output of \code{\link[=PHM]{PHM()}}}

\item{threshold}{\eqn{P_{\rm{mc}}} threshold, default is 0.01}
}
\value{
Result of the PHM merging procedure terminated when the \eqn{P_{\rm{mc}}} threshold is satisfied
export
}
\description{
Each step of the PHM algorithm reduces \eqn{P_{\rm{mc}}}. This gives the results from the PHM algorithm terminated when \eqn{P_{\rm{mc}}} falls below some specified threshold.
}
